fix: media id type from int to long

+++ src/main/java/com/codecool/poster/repository/MediaRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Collection;
 
 @Repository
-public interface MediaRepository extends JpaRepository<Media, Integer> {
+public interface MediaRepository extends JpaRepository<Media, Long> {
     Collection<Media> findAllByPostId(long id);
 }
